feat: squad feature access (#1511)
Created a new hook for opening the squad modal (this works based on notifications)
WT-924 #done
