+++ packages/shared/src/components/sidebar/Sidebar.tsx
@@ -1,10 +1,4 @@
-import React, {
-  ReactElement,
-  useContext,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { ReactElement, useContext, useMemo, useState } from 'react';
 import classNames from 'classnames';
 import dynamic from 'next/dynamic';
 import SettingsContext from '../../contexts/SettingsContext';
@@ -34,6 +28,7 @@ import { ProfilePicture } from '../ProfilePicture';
 import { useLazyModal } from '../../hooks/useLazyModal';
 import { LazyModal } from '../modals/common/types';
 import { Squad } from '../../graphql/squads';
+import { useCreateSquadModal } from '../../hooks/useCreateSquadModal';
 
 const UserSettingsModal = dynamic(
   () =>
@@ -63,6 +58,7 @@ export default function Sidebar({
     loadedSettings,
     optOutWeeklyGoal,
   } = useContext(SettingsContext);
+  const { openSquadBetaModal } = useCreateSquadModal();
   const { openModal } = useLazyModal();
   const [showSettings, setShowSettings] = useState(false);
   const {
@@ -84,24 +80,6 @@ export default function Sidebar({
     state: openMobileSidebar,
     action: setOpenMobileSidebar,
   });
-
-  const previousRef = useRef(null);
-
-  const openNewSquadModal = () =>
-    openModal({
-      type: LazyModal.NewSquad,
-      props: {
-        onPreviousState: () => previousRef.current(),
-      },
-    });
-  const openSquadBetaModal = () =>
-    openModal({
-      type: LazyModal.BetaSquad,
-      props: {
-        onNext: openNewSquadModal,
-      },
-    });
-  previousRef.current = openSquadBetaModal;
 
   const openLockedSquadModal = (squad: Squad) => {
     openModal({
